Add EBS io2 support and fixes

+++ scripts/install-master.sh
@@ -78,14 +78,24 @@ create_volume () {
 
 
 set_noop_scheduler () {
-	log `date` "Setting i/o scheduler to noop for each physical volume"
 	for i in `pvs | grep dev | awk '{print $1}' | sed s/\\\/dev\\\///`
 	do
-	  echo "noop" > /sys/block/$i/queue/scheduler
-	  printf "$i: "
-	  cat /sys/block/$i/queue/scheduler
-	done
-
+	  if [[ "$i" != *"nvme"* ]]; then
+	    if [[ $(grep -wq noop /sys/block/$i/queue/scheduler; echo $?) -eq 0 ]]; then
+	      log `date` "Setting i/o scheduler to noop for each physical volume"
+	      echo "noop" > /sys/block/$i/queue/scheduler
+	      printf "$i: "
+	      cat /sys/block/$i/queue/scheduler
+	    else
+	      log `date` "Setting i/o scheduler to none for each physical volume"
+	      echo "none" > /sys/block/$i/queue/scheduler
+	      printf "$i: "
+	      cat /sys/block/$i/queue/scheduler
+	    fi
+      else
+        log `date` "Skipping IO block scheduler change - NVMe device detected"
+      fi
+    done
 }
